separated Header component/new data folder/tangible routing done
