+++ src/SpellTypePage.js
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react'
-import spellbook from './Spellbook'
+import spellbook from './data/Spellbook'
 import { useParams } from 'react-router-dom'
 import './App.css'
 
@@ -13,6 +13,7 @@ function SpellTypePage () {
 
   console.log(type)
   const selectedSpells = spellbook.filter(spell => spell.type === type)
+
   useEffect(() => {
     document.title = `${type.toUpperCase()} - LazyPace's Book of Spells`
   }, [type])
@@ -37,7 +38,7 @@ function SpellTypePage () {
                 <td className={spell.restricted ? 'restricted-spell' : ''}>
                   <em> {spell.incantation || 'No standard incantation'}</em>
                 </td>
-                <td>{spell.nickname || 'N/A'}</td>
+                <td>{spell.nickname || '-'}</td>
                 <td>{spell.effect}</td>
               </tr>
             ))}
