+++ src/TangibleTypePage.js
@@ -1,12 +1,19 @@
 import React, { useEffect } from 'react'
-import potionbook from './Potionbook'
+import potionbook from './data/Potionbook'
 import { useParams } from 'react-router-dom'
 import './App.css'
 
 function TangibleTypePage () {
   const { type } = useParams()
+  const { tangibles } = potionbook
 
-  console.log(type)
+  console.log('the type is ', type)
+
+  return (
+    <div className={`tangible-header`}>
+      <h2>{type.toUpperCase()}</h2>
+    </div>
+  )
 }
 
 export default TangibleTypePage
